# opt/props: use EquivGroups in FuncDepSet for tracking equivalences