--- a/pkg/sql/opt/memo/testdata/logprops/delete
+++ b/pkg/sql/opt/memo/testdata/logprops/delete
@@ -152,7 +152,7 @@ DELETE FROM abcde WHERE b=c RETURNING *;
 project
  ├── columns: a:1(int!null) b:2(int!null) c:3(int!null) d:4(int)
  ├── volatile, mutations
- ├── fd: (2)==(3), (3)==(2), (2)-->(4)
+ ├── fd: (2)-->(4), (2)==(3), (3)==(2)
  ├── prune: (1-4)
  └── delete abcde
       ├── columns: a:1(int!null) b:2(int!null) c:3(int!null) d:4(int) rowid:5(int!null)
@@ -165,7 +165,7 @@ project
       │    └── rowid:13 => rowid:5
       ├── volatile, mutations
       ├── key: (5)
-      ├── fd: (2)==(3), (3)==(2), (5)-->(1-4), (2)-->(4)
+      ├── fd: (5)-->(1-4), (2)-->(4), (2)==(3), (3)==(2)
       └── select
            ├── columns: a:9(int!null) b:10(int!null) c:11(int!null) d:12(int) rowid:13(int!null) e:14(int) crdb_internal_mvcc_timestamp:15(decimal) tableoid:16(oid)
            ├── key: (13)