--- a/pkg/sql/opt/memo/testdata/logprops/max1row
+++ b/pkg/sql/opt/memo/testdata/logprops/max1row
@@ -63,7 +63,7 @@ select
  ├── columns: a:1(int!null) b:2(int!null) c:3(int!null)
  ├── cardinality: [0 - 1]
  ├── key: ()
- ├── fd: ()-->(1-3), (3)==(2), (2)==(3)
+ ├── fd: ()-->(1-3), (2)==(3), (3)==(2)
  ├── scan abc
  │    ├── columns: a:1(int!null) b:2(int) c:3(int)
  │    ├── key: (1)