--- a/pkg/sql/opt/memo/testdata/logprops/scan
+++ b/pkg/sql/opt/memo/testdata/logprops/scan
@@ -422,13 +422,13 @@ SELECT * FROM computed
 project
  ├── columns: i:1(int!null) s:2(string) d:3(decimal) c_i_expr:4(string!null) c_s:5(string) c_d:6(decimal) c_d_expr:7(string)
  ├── key: (1)
- ├── fd: (1,4)-->(2,3,5-7), (3,7)~~>(1,2,4-6), (1)-->(4), (3)==(6), (6)==(3), (2)~~>(1,3-7), (2)==(5), (5)==(2)
+ ├── fd: (1,4)-->(2,3,5-7), (3,7)~~>(1,2,4-6), (1)-->(4), (2)~~>(1,3-7), (3)==(6), (6)==(3), (2)==(5), (5)==(2)
  ├── prune: (1-7)
  ├── interesting orderings: (+4,+1) (+7,+(3|6),+4,+1)
  └── project
       ├── columns: c_s:5(string) i:1(int!null) s:2(string) d:3(decimal) c_i_expr:4(string!null) c_d:6(decimal) c_d_expr:7(string) crdb_internal_mvcc_timestamp:8(decimal) tableoid:9(oid)
       ├── key: (1)
-      ├── fd: (1,4)-->(2,3,6-9), (3,7)~~>(1,2,4,6,8,9), (1)-->(4), (3)==(6), (6)==(3), (2)~~>(1,3,4,6-9), (2)==(5), (5)==(2)
+      ├── fd: (1,4)-->(2,3,6-9), (3,7)~~>(1,2,4,6,8,9), (1)-->(4), (2)~~>(1,3,4,6-9), (3)==(6), (6)==(3), (2)==(5), (5)==(2)
       ├── prune: (1-9)
       ├── interesting orderings: (+4,+1) (+7,+(3|6),+4,+1)
       ├── scan computed
@@ -451,7 +451,7 @@ project
       │    │         └── cast: STRING [type=string]
       │    │              └── variable: d:3 [type=decimal]
       │    ├── key: (1)
-      │    ├── fd: (1,4)-->(2,3,6-9), (3,7)~~>(1,2,4,6,8,9), (1)-->(4), (3)==(6), (6)==(3), (2)~~>(1,3,4,6-9)
+      │    ├── fd: (1,4)-->(2,3,6-9), (3,7)~~>(1,2,4,6,8,9), (1)-->(4), (2)~~>(1,3,4,6-9), (3)==(6), (6)==(3)
       │    ├── prune: (1-4,6-9)
       │    └── interesting orderings: (+4,+1) (+7,+(3|6),+4,+1)
       └── projections