… pkg/ccl/logictestccl/testdata/logic_test/regional_by_row_query_behavior
… pkg/sql/opt/memo/testdata/logprops/join
… pkg/sql/opt/memo/testdata/stats_quality/tpch/q02
… pkg/sql/opt/memo/testdata/stats_quality/tpch/q05
… pkg/sql/opt/memo/testdata/stats_quality/tpch/q08
… pkg/sql/opt/memo/testdata/stats_quality/tpch/q10
… pkg/sql/opt/memo/testdata/stats_quality/tpch/q19
… pkg/sql/opt/norm/rules/join.opt
… pkg/sql/opt/norm/rules/select.opt
… pkg/sql/opt/norm/testdata/rules/combo
… pkg/sql/opt/norm/testdata/rules/decorrelate
… pkg/sql/opt/norm/testdata/rules/inline
… pkg/sql/opt/norm/testdata/rules/join
… pkg/sql/opt/norm/testdata/rules/limit
… pkg/sql/opt/norm/testdata/rules/prune_cols
… pkg/sql/opt/norm/testdata/rules/reject_nulls
… pkg/sql/opt/norm/testdata/rules/scalar
… pkg/sql/opt/norm/testdata/rules/select
… pkg/sql/opt/norm/testdata/rules/with
… pkg/sql/opt/props/BUILD.bazel
… pkg/sql/opt/props/equiv_set.go
… pkg/sql/opt/props/equiv_set_test.go
… pkg/sql/opt/props/func_dep.go
… pkg/sql/opt/props/func_dep_rand_test.go
… pkg/sql/opt/props/func_dep_test.go
… pkg/sql/opt/xform/join_order_builder.go
… pkg/sql/opt/xform/testdata/coster/join
… pkg/sql/opt/xform/testdata/external/hibernate
… pkg/sql/opt/xform/testdata/external/liquibase
… pkg/sql/opt/xform/testdata/external/navicat