#### props: don't leave unmerged equiv groups in EquivSet

This commit fixes a bug in `EquivSet.tryMergeGroups`, which could cause
an invariant (that equiv groups are non-intersecting) to be violated.
This would potentially cause re-ordered joins to have redundant equality
filters, and prevent join filter push-down in rare cases. This commit
also adds verification to the set for test builds, and a unit test for
the `Add` method.

Fixes #137381

Release note: None

#### props: rename EquivSet to EquivGroups

This commit renames `EquivSet` to `EquivGroups` to better reflect the
fact that it tracks _groups_ of equivalent groups, rather than a single
set of equivalent columns.

Epic: None

Release note: None

#### props: remove initial buffer from EquivGroups

This commit removes the inline buffer from `EquivGroups`, as well as the
`NewEquivGroups()` that was previously used to set up the buffer. This
simplifies usage, and doesn't appear to affect performance when the
set is used in `JoinOrderBuilder` or in `FuncDepSet`.

Informs #83963

Release note: None

#### props: add methods to EquivGroups for use in FuncDepSet

This commit adds several new methods along with unit tests to `EquivGroups`
to prepare its use in tracking equivalencies in `FuncDepSet`.

Informs #83963

Release note: None

#### opt/props: use EquivGroups in FuncDepSet for tracking equivalences

Previously, `FuncDepSet` was fairly wasteful in how it tracked sets of
equivalent columns: for each column in the equiv group, an FD was maintained
from that column to all other columns in the group. This meant that there
were `2n` `ColSets` for each equiv group (where `n` is the number of columns
in the group).

This patch modifies `FuncDepSet` and its internals to use `props.EquivGroups`
instead, which keeps a single `ColSet` for each equiv group. This significantly
cuts down on allocations for queries with many columns and equalities, both
because less `ColSets` spill to heap, and because less FDs are added to
the `deps` slice.

Fixes #83963

Release note: None