@DrewKimball 2024-12-17
Continuation of https://github.com/cockroachdb/cockroach/pull/117272. First commit is #137558. Here's an updated slow-queries benchmark against master:
```
name                                       old time/op    new time/op    delta
SlowQueries/slow-query-1/reorder-join-0-8    1.03ms ± 1%    0.91ms ± 4%  -11.62%  (p=0.000 n=9+10)
SlowQueries/slow-query-1/reorder-join-8-8    13.5ms ± 0%    11.7ms ± 0%  -13.53%  (p=0.000 n=9+10)
SlowQueries/slow-query-2/reorder-join-0-8    1.97ms ± 1%    1.69ms ± 1%  -14.50%  (p=0.000 n=9+9)
SlowQueries/slow-query-2/reorder-join-8-8     217ms ± 2%     204ms ± 2%   -6.00%  (p=0.000 n=10+10)
SlowQueries/slow-query-3/reorder-join-0-8    43.0ms ± 1%    43.0ms ± 2%     ~     (p=1.000 n=10+10)
SlowQueries/slow-query-3/reorder-join-8-8    45.6ms ± 1%    45.1ms ± 2%   -1.10%  (p=0.023 n=10+10)
SlowQueries/slow-query-4/reorder-join-0-8    60.9ms ± 0%     6.4ms ± 1%  -89.45%  (p=0.000 n=8+9)
SlowQueries/slow-query-4/reorder-join-8-8    61.2ms ± 0%     6.4ms ± 0%  -89.50%  (p=0.000 n=8+8)
SlowQueries/slow-query-5/reorder-join-0-8    40.9ms ± 0%    34.9ms ± 1%  -14.80%  (p=0.000 n=8+9)
SlowQueries/slow-query-5/reorder-join-8-8     724ms ± 1%     668ms ± 0%   -7.77%  (p=0.000 n=10+8)
SlowQueries/slow-query-6/reorder-join-0-8    24.7ms ± 1%    14.7ms ± 1%  -40.66%  (p=0.000 n=9+10)
SlowQueries/slow-query-6/reorder-join-8-8     990ms ± 5%     776ms ± 2%  -21.60%  (p=0.000 n=10+10)
SlowQueries/slow-query-7/reorder-join-0-8    34.4ms ± 1%    19.4ms ± 0%  -43.62%  (p=0.000 n=9+10)
SlowQueries/slow-query-7/reorder-join-8-8     1.25s ± 1%     0.96s ± 1%  -23.67%  (p=0.000 n=9+8)

name                                       old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1/reorder-join-0-8     690kB ± 0%     571kB ± 0%  -17.21%  (p=0.000 n=8+9)
SlowQueries/slow-query-1/reorder-join-8-8    6.66MB ± 0%    5.93MB ± 0%  -10.94%  (p=0.000 n=8+7)
SlowQueries/slow-query-2/reorder-join-0-8     957kB ± 0%     816kB ± 0%  -14.73%  (p=0.000 n=9+10)
SlowQueries/slow-query-2/reorder-join-8-8    51.1MB ± 0%    52.9MB ± 0%   +3.48%  (p=0.000 n=10+10)
SlowQueries/slow-query-3/reorder-join-0-8    39.9MB ± 0%    39.7MB ± 0%   -0.39%  (p=0.000 n=10+8)
SlowQueries/slow-query-3/reorder-join-8-8    42.0MB ± 0%    41.7MB ± 0%   -0.88%  (p=0.000 n=10+10)
SlowQueries/slow-query-4/reorder-join-0-8    30.3MB ± 0%     4.1MB ± 0%  -86.50%  (p=0.000 n=10+10)
SlowQueries/slow-query-4/reorder-join-8-8    30.3MB ± 0%     4.1MB ± 0%  -86.50%  (p=0.000 n=8+8)
SlowQueries/slow-query-5/reorder-join-0-8    33.4MB ± 0%    32.3MB ± 0%   -3.14%  (p=0.000 n=9+10)
SlowQueries/slow-query-5/reorder-join-8-8     256MB ± 0%     249MB ± 0%   -2.62%  (p=0.000 n=10+10)
SlowQueries/slow-query-6/reorder-join-0-8    10.5MB ± 0%     7.7MB ± 0%  -27.12%  (p=0.000 n=9+10)
SlowQueries/slow-query-6/reorder-join-8-8     424MB ± 1%     338MB ± 0%  -20.08%  (p=0.000 n=10+10)
SlowQueries/slow-query-7/reorder-join-0-8    14.0MB ± 0%    10.2MB ± 0%  -27.24%  (p=0.000 n=8+8)
SlowQueries/slow-query-7/reorder-join-8-8     534MB ± 0%     409MB ± 0%  -23.39%  (p=0.000 n=10+8)

name                                       old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1/reorder-join-0-8     3.86k ± 0%     3.25k ± 0%  -15.79%  (p=0.000 n=8+9)
SlowQueries/slow-query-1/reorder-join-8-8     65.7k ± 0%     52.5k ± 0%  -20.06%  (p=0.000 n=8+7)
SlowQueries/slow-query-2/reorder-join-0-8     3.96k ± 0%     3.85k ± 0%   -2.63%  (p=0.000 n=9+10)
SlowQueries/slow-query-2/reorder-join-8-8      265k ± 0%      271k ± 0%   +2.36%  (p=0.000 n=10+10)
SlowQueries/slow-query-3/reorder-join-0-8      300k ± 0%      296k ± 0%   -1.07%  (p=0.000 n=10+8)
SlowQueries/slow-query-3/reorder-join-8-8      320k ± 0%      312k ± 0%   -2.41%  (p=0.000 n=10+10)
SlowQueries/slow-query-4/reorder-join-0-8      166k ± 0%       28k ± 0%  -83.02%  (p=0.000 n=10+10)
SlowQueries/slow-query-4/reorder-join-8-8      166k ± 0%       28k ± 0%  -83.02%  (p=0.000 n=8+8)
SlowQueries/slow-query-5/reorder-join-0-8      210k ± 0%      200k ± 0%   -5.00%  (p=0.000 n=9+10)
SlowQueries/slow-query-5/reorder-join-8-8     4.43M ± 0%     4.38M ± 0%   -1.17%  (p=0.000 n=10+10)
SlowQueries/slow-query-6/reorder-join-0-8      110k ± 0%       76k ± 0%  -31.18%  (p=0.000 n=8+9)
SlowQueries/slow-query-6/reorder-join-8-8     6.90M ± 1%     5.46M ± 0%  -20.84%  (p=0.000 n=10+9)
SlowQueries/slow-query-7/reorder-join-0-8      175k ± 0%      116k ± 0%  -33.65%  (p=0.000 n=8+8)
SlowQueries/slow-query-7/reorder-join-8-8     10.2M ± 0%      7.7M ± 0%  -24.02%  (p=0.000 n=10+8)
```